SUBMIT-587bMoves SubmissionItem Cypress test to staff directory (#650)
* Adds SubmissionItem Cypress component test

Adds a Cypress component test for the SubmissionItem route.

This test verifies that the page renders correctly and that the review section is visible. It also includes a test case for when the consultation check is completed, ensuring that the relevant notification is displayed.

Also moves the test constants to separate file for reusability.

* Moves SubmissionItem to staff directory

Relocates the SubmissionItem component to the _staff directory
to better organize staff-specific components.

Addresses issue of component placement within the project structure.

Fixes #587

* Uses staff account for internal documents

Ensures that the tests for internal documents use a staff account mock, reflecting the intended user role for this functionality.
This change aligns the tests with the actual user authentication context required for accessing and managing internal staff documents.

* Moves internal documents component

Moves the internal documents component to the staff directory.