SUBMIT-587 Refactors and adds Cypress component tests (#654)
* Adds SubmissionItem Cypress component test

Adds a Cypress component test for the SubmissionItem route.

This test verifies that the page renders correctly and that the review section is visible. It also includes a test case for when the consultation check is completed, ensuring that the relevant notification is displayed.

Also moves the test constants to separate file for reusability.

* Moves SubmissionItem to staff directory

Relocates the SubmissionItem component to the _staff directory
to better organize staff-specific components.

Addresses issue of component placement within the project structure.

Fixes #587

* Uses staff account for internal documents

Ensures that the tests for internal documents use a staff account mock, reflecting the intended user role for this functionality.
This change aligns the tests with the actual user authentication context required for accessing and managing internal staff documents.

* Moves internal documents component

Moves the internal documents component to the staff directory.

* Adds staff user authentication roles

Adds mock authentication roles for staff users in the consultation record component test.

This allows the test to properly simulate a staff user with the appropriate permissions, which is necessary to test the intended functionality of the review section.

* Test review section for staff who is not manager

* Renames and updates consultation record test

Renames the submission item test file to more accurately reflect its purpose, focusing on consultation record functionality.

Updates the test suite to distinguish between manager and staff roles when testing the review section.

* Refactors Cypress tests for better structure

Moves mock constants to a shared utility directory.
Creates a helper function for mounting pages in Cypress tests to reduce duplication.
Refactors consultation record tests to use new mounting helper and mock constants.

Addresses: SUBMIT-task#587

* Removes obsolete screenshot file.

Deletes a failed screenshot artifact that is no longer relevant.

* Removes obsolete SubmissionItem test file

Deletes the SubmissionItem.cy.tsx file, which is no longer needed.

* Adds consultation record tests

Adds Cypress tests for the consultation record page for both proponent and staff views.

These tests verify that the page renders correctly and displays the associated documents.
It also mocks account and router to pass the authentication context needed.

* Remove console logs