Replaces global loader with component-level backdrop (#681)
* Replaces Zustand loader with component prop

Removes the global Zustand store for the loader backdrop and instead passes an `isOpen` prop to the `SubmitLoaderBackdrop` component. This allows for more localized control of the backdrop's visibility within each component that uses it, simplifying state management and improving component reusability.

* Removes loader backdrop provider

Removes the loader backdrop provider from the application's main component.

The overlay functionality is no longer needed as a global provider.