Adds endpoint to retrieve current user (#697)
* Adds endpoint to retrieve current user

Implements a new endpoint to retrieve the current authenticated user.
This endpoint also handles auto-provisioning of staff users if they have valid roles but lack a corresponding database record.
The frontend is updated to call this new endpoint.

* Implements user creation on authentication

Updates the user endpoint to create a new user if one doesn't exist,
instead of just retrieving the user.
This ensures that every authenticated user has a corresponding
record in the database.

Related to CENTRE-task#59

* Fix no blank lines after function docstring